requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
oauth2client==4.1.3
protobuf==3.15.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
python-jwt==3.3.0
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
touch==2020.12.3
urllib3==1.26.3
values==2020.12.3
Werkzeug==1.0.1
